feat(discount): change discount type to float64 for improved precision

+++ internal/logic/public/order/getDiscount.go
@@ -3,7 +3,7 @@ package order
 import "github.com/perfect-panel/server/internal/types"
 
 func getDiscount(discounts []types.SubscribeDiscount, inputMonths int64) float64 {
-	var finalDiscount int64 = 100
+	var finalDiscount float64 = 100
 
 	for _, discount := range discounts {
 		if inputMonths >= discount.Quantity && discount.Discount < finalDiscount {
@@ -11,5 +11,5 @@ func getDiscount(discounts []types.SubscribeDiscount, inputMonths int64) float64
 		}
 	}
 
-	return float64(finalDiscount) / float64(100)
+	return finalDiscount / float64(100)
 }
